requirements.txt
pyserial==3.5
matplotlib==3.8.2
scipy==1.13.0
numpy==1.26.4